feat: Add a window_handle_any_thread function
By making the `window_handle()` function only work on the origin thread
for Windows, we prevent thread-safe usage of the window handle. This is
intended but might not be practical for applications that are sure that
they only use thread-safe APIs.

Therefore in this commit I add a `window_handle_any_thread` function
that allows for using the window handle on other threads.

I should also add a newtype wrapper around `Window` that calls this
function, in order for it to be used as a `HasWindowHandle` impl.

Signed-off-by: John Nunley <dev@notgull.net>